poc_tests/ocr_validation/requirements.txt
google-cloud-vision==3.8.0
google-cloud-aiplatform==1.60.0
PyPDF2==3.0.1
Pillow==10.2.0
opencv-python==4.9.0.80
python-dotenv==1.0.1
pytest==8.0.0
requests==2.31.0